# std: Optimize Vec::from_iter