@alexcrichton 2015-02-12
cc @Gankro, https://github.com/rust-lang/rfcs/pull/832