@huonw 2015-02-12
`33 ns/iter` seems rather short, even for a memcpy/memset?