@huonw 2015-02-12
@bors r+ 2543 rollup