@alexcrichton 2015-02-12
@bors: retry