@alexcrichton 2015-02-12
@bors: r+ 985fc7d