@alexcrichton 2015-02-12
@bors: r=huonw 985fc7d